copies: narrow the reference-index grids i32 -> i8 (-103 MB of memset)
First win from the decoder copy hunt, and the instrument mattered more
than the fix.

THE INSTRUMENT. A source grep found 101 copy_from_slice sites in the
decoder and ranked nothing -- b_skip_slow's 384-byte hop and the nnz
grid's 4-byte one look identical in a grep, and one is on the rare path.
Added bench/copy_census.sh, which censuses the EMITTED ASSEMBLY for real
call memcpy/memset/memmove: 623 calls across 103 symbols, i.e. the ones
with call overhead attached, including the ones the optimiser introduced
and excluding the ones LLVM already inlined away.

WHAT THAT RULED OUT, which is most of the value:
  * All 5 drain() sites are full drains, not front-drains. The O(n^2)
    pattern this skill greps for first is NOT present.
  * The census's top three symbols (80 copies in the frame-MT sink, 21 in
    Sender<JobOut>, 17 in Sender<PixelCtx>) execute ZERO times in a
    single-threaded decode -- confirmed by a runtime counter reading
    jobs=0 rows=0 needctx=0. EdcMsg is 40-80 bytes anyway; the 2.7 KB
    jobs are already pooled behind a Box.
  * recon_p_inter_parts stages MC through a 256-byte local. The stack
    scratch is NOT memset (checked the .s for an actual callq memset --
    LLVM elided it), so the standing "2 KB stack temporary" lever does
    not apply here.

WHAT IT FOUND. A per-picture counter on `refill`, the single funnel for
the pool re-arm, with #[track_caller] to bucket by call site:

  733,764,000 bytes of memset over a 60-frame 720p decode -- 12.2 MB per
  picture, ~7.8% of decode at memset bandwidth. Four grids were 56% of
  it, and the per-call sizes identify them exactly:
    mv_y      450 KB/call   138 MB
    mv1       450 KB/call   138 MB
    ref_idx_y 225 KB/call    69 MB
    ref_idx1  225 KB/call    69 MB

THE FIX. A ref_idx is -1..31 by the spec (7.4.5.1: num_ref_idx_active is
at most 32), so three of every four bytes were padding. Narrowed the two
reference grids, BlockInfo::ref_id/ref_id1, the encoder's matching grids
and its deblock scratch, and the coloc metadata that mirrors them.
MvField (a public diagnostic type) is widened at the clone instead, and
readers that want the POC/NO_REF space use i32::from, which sign-extends
so a stored -1 stays -1 while an ABSENT entry still yields NO_REF.

MEASURED: 733,764,000 -> 630,084,000 B, i.e. -103,680,000 (-14.1%) of
the pool clear, matching the prediction exactly. The cache half is the
part the census cannot see: these grids are read for every neighbour
lookup in motion prediction, and they are now a quarter of the size.

Gates: 68/68 ALL BYTE-IDENTICAL (encoder AND decoder touched, so this was
mandatory); full workspace tests green; clippy clean under -D warnings.

Still open, same census: mv_y/mv1 are (i32,i32) and 276 MB of the
remaining 630 MB. pack_mb already narrows them to i16 unconditionally,
so the width is provably sufficient -- that is the next item.

+++ crates/rusty_h264-decoder/src/lib.rs
@@ -480,7 +480,9 @@ pub(crate) struct RefFrame {
     /// for intra), and the block-grid width. Read as the *co-located* picture's
     /// motion for B-slice direct prediction (`colZeroFlag`, temporal direct).
     pub mv: Vec<(i32, i32)>,
-    pub ref_idx: Vec<i32>,
+    /// Narrowed with the grid it mirrors (a `ref_idx` is -1..31; see
+    /// `FrameDecoder::ref_idx_y`).
+    pub ref_idx: Vec<i8>,
     /// Per-4×4-block **List-1** motion. Needed because the co-located motion
     /// derivation (spec §8.4.1.2.1) falls back to List-1 when the co-located block
     /// has no List-0 prediction (`predFlagL0Col == 0`). A co-located picture only
@@ -488,7 +490,7 @@ pub(crate) struct RefFrame {
     /// what b-pyramid produces, so this stayed unexercised until B-references
     /// appeared.
     pub mv1: Vec<(i32, i32)>,
-    pub ref_idx1: Vec<i32>,
+    pub ref_idx1: Vec<i8>,
     /// Per-4×4-block POC of the List-0 picture each block referenced (`i32::MIN`
     /// for intra). Used by temporal direct's `MapColToList0` (the co-located
     /// reference index alone is meaningless in the current list).
@@ -570,9 +572,11 @@ pub(crate) struct LiveMeta {
     pub long_term: bool,
     pub long_term_idx: u32,
     pub mv: Vec<(i32, i32)>,
-    pub ref_idx: Vec<i32>,
+    /// Narrowed with the grid it mirrors (a `ref_idx` is -1..31; see
+    /// `FrameDecoder::ref_idx_y`).
+    pub ref_idx: Vec<i8>,
     pub mv1: Vec<(i32, i32)>,
-    pub ref_idx1: Vec<i32>,
+    pub ref_idx1: Vec<i8>,
     pub ref_poc: Vec<i32>,
     pub w4: usize,
     /// True once finalize has published coloc motion (temporal direct may read).
