+++ crates/rusty_h264-encoder/src/lib.rs
@@ -566,14 +566,16 @@ pub(crate) struct RefFrame {
     /// read as the co-located picture (`RefPicList1[0]`) when deriving a B-frame's
     /// spatial-direct `colZeroFlag`. `ref_idx == -1` marks intra/uncoded blocks.
     pub mv: Vec<(i32, i32)>,
-    pub ref_idx: Vec<i32>,
+    /// Narrowed with the grid it mirrors (a `ref_idx` is -1..31; the deblock
+    /// `BlockInfo::ref_id` is `&[i8]`).
+    pub ref_idx: Vec<i8>,
     /// List-1 motion of the picture (b-pyramid: a REFERENCE B can be the
     /// co-located picture, and its L1-only blocks read from here — the exact
     /// List-1 colZeroFlag defect the decoder already root-caused and fixed;
     /// the encoder's direct derivation must mirror it or pyramid recon
     /// drifts). EMPTY for P/I references (no List 1 exists there).
     pub mv1: Vec<(i32, i32)>,
-    pub ref_idx1: Vec<i32>,
+    pub ref_idx1: Vec<i8>,
     /// Blocks-wide (`mb_w*4`), so the co-located index is `by*w4 + bx`.
     pub w4: usize,
     /// Cached half-pel luma planes, built on first sub-pel motion-search use.
